+++ projects/isy-angular-widgets/src/lib/hauptfenster/hauptfenster.module.ts
@@ -3,10 +3,11 @@ import {HauptfensterComponent} from './hauptfenster.component';
 import {ButtonModule} from 'primeng/button';
 import {CommonModule} from '@angular/common';
 import {MegaMenuModule} from 'primeng/megamenu';
+import {SeitentoolbarComponent} from './components/seitentoolbar/seitentoolbar.component';
 
 @NgModule({
   declarations: [HauptfensterComponent],
-  imports: [ButtonModule, CommonModule, MegaMenuModule],
-  exports: [HauptfensterComponent]
+  exports: [HauptfensterComponent],
+  imports: [ButtonModule, CommonModule, MegaMenuModule, SeitentoolbarComponent]
 })
 export class HauptfensterModule {}